Se agrega la variable puerto en la configuracion de la base de datos

+++ controllers/actaController.js
@@ -102,148 +102,7 @@ module.exports = {
 		}
 	},
 
-	async updateX(req, res){
-		try{
-			let numeroActivos = 0;
-			let totalValor = 0;
-			let responsePromises = [];
-			if(req.body.Detalle != undefined && req.body.Detalle.length > 0){
-				
-				// await req.body.Detalle.forEach( async(element) => {
-		  //         if (element.estado == "A"){
-		  //         	//Registro ya existe, solo acumula totales
-		  //           numeroActivos++;
-		  //           totalValor += element.valorCompraIva;
-		  //         }
-		  //         if (element.estado == "I"){
-		  //           //Registro nuevo, se procede a insertar el detalle
-		  //           let detalle = {
-				// 		actaId: element.actaId,
-			 //            activoId: element.activoId,
-			 //            codigo: element.codigo,
-			 //            codigoConcesionaria: element.codigoConcesionaria,
-			 //            usuarioIngresa: element.usuarioIngresa,
-			 //            fechaIngresa : element.fechaIngresa,
-			 //            estado: "A",
-			 //            fechaUltimoCambio : element.fechaUltimoCambio,
-			 //            valorCompraIva: element.valorCompraIva
-		  //           }
-		  //  //          let promise = ActaDetalle.create(detalle)
-				// 	// 	.then(data => {
-				// 	// 		numeroActivos++;
-		  //  //          		totalValor += element.valorCompraIva;
-		  //  //          		logger.error('Actualiza '+element.id);
-		  //  //          		logger.error(numeroActivos + ' ' + totalValor);
-				// 	// 	})
-				// 	// 	.catch(err => res.json({error: err}));
-				// 	// responsePromises.push(promise);
-				// 	//(async() => await ActaDetalle.create(detalle));
-				// 	//responsePromises.push(promise);
-				// 	let res= await procesaDetalle(detalle);
-				// 		numeroActivos++;
-	   //          		totalValor += element.valorCompraIva;
-	   //          		logger.error('Actualiza '+element.id);
-	   //          		logger.error(numeroActivos + ' ' + totalValor);
-
-
-		  //         }
-		  //         if (element.estado == "X"){
-		  //           //Registro a eliminar
-		  //           try{
-		  //             let promise = ActaDetalle.destroy({where: {id: element.id}})
-
-		  //           }catch(err){next(err);}
-		  //         }
-		  //       });
-
-		  await req.body.Detalle.forEach( async(element) => {
-		  	if (element.estado == "I"){
-		            //Registro nuevo, se procede a insertar el detalle
-		            let detalle = {
-		            	id: null,
-						actaId: element.actaId,
-			            activoId: element.activoId,
-			            codigo: element.codigo,
-			            codigoConcesionaria: element.codigoConcesionaria,
-			            usuarioIngresa: element.usuarioIngresa,
-			            fechaIngresa : element.fechaIngresa,
-			            estado: "A",
-			            fechaUltimoCambio : element.fechaUltimoCambio,
-			            valorCompraIva: element.valorCompraIva
-		            }
-		            const res= await procesaDetalle(detalle);
-		            logger.error(res);
-		            numeroActivos++;
-		         		totalValor += element.valorCompraIva;
-		          		logger.error('Actualiza '+detalle.id);
-		         		logger.error(numeroActivos + ' ' + totalValor);
-		        }
-		  	// 	const resultadoDetalle = await Promise.all(req.body.Detalle.map(element => {
-		  	// 		if (element.estado == "I"){
-		   //          //Registro nuevo, se procede a insertar el detalle
-		   //          let detalle = {
-					// 	actaId: element.actaId,
-			  //           activoId: element.activoId,
-			  //           codigo: element.codigo,
-			  //           codigoConcesionaria: element.codigoConcesionaria,
-			  //           usuarioIngresa: element.usuarioIngresa,
-			  //           fechaIngresa : element.fechaIngresa,
-			  //           estado: "A",
-			  //           fechaUltimoCambio : element.fechaUltimoCambio,
-			  //           valorCompraIva: element.valorCompraIva
-		   //          }
-		   //          ActaDetalle.create(detalle)
-					// .then(data => {
-					// 	numeroActivos++;
-		   //       		totalValor += element.valorCompraIva;
-		   //        		logger.error('Actualiza '+detalle.id);
-		   //       		logger.error(numeroActivos + ' ' + totalValor);
-					// })
-					// .catch(err => res.json({error: err}));
-		  	// 	}}));
-			});
-		}
-			logger.error(responsePromises);
-			logger.error('resultadoDetalle')
-			//logger.error(resultadoDetalle);
-			//let responses = await Promise.all(responsePromises);
-
-			logger.error(numeroActivos + ' ' + totalValor);
-			logger.error('update acta');
-			let record = {
-				fechaActa: req.body.fechaActa,
-				numeroActa: req.body.numeroActa,
-				comentarios: req.body.comentarios,
-				estadoInicialId: req.body.estadoInicialId,
-				estadoFinalId: req.body.estadoFinalId,
-				usuarioAprueba: req.body.usuarioAprueba,
-				fechaAprueba: req.body.fechaAprueba,
-				numeroActivos: numeroActivos,
-				totalValor: totalValor,
-				estado: req.body.estado
-			}
-			logger.error('Actualizar acta ');
-			Acta.update(record, {
-				where: {id: req.body.id}
-			})
-				.then(data => {
-					res.status(200).json({success:true, message: "Acta actualizada con exito", respuesta: data});
-				}).catch(err => res.json({error: err}));
-		}catch(e){
-			logger.error(e);
-			res.send(500);
-		}
-	},
-
 	async update(req, res){		
-/* 		let query = "begin tmp_retorno(); end;";
-		const[result, metadata] = await db.sequelize.query(query).then().catch(err => logger.error(err));
-		logger.error('resp sp');
-		logger.error(JSON.stringify(result) +" metadata " + JSON.stringify(metadata));
-		logger.error(metadata);
- */		
-
-
 		let transaction;
 		try{
 			transaction =  await db.sequelize.transaction();
@@ -281,16 +140,8 @@ module.exports = {
 						}catch(err) {
 							logger.error(err);
 							throw new Error('Error en creacion de detalle');
-							//res.status(501).json({error: err}).send();
 						}
-/* 						const eje = await ActaDetalle.create(detalle
-						).then(data => {
-			            	numeroActivos++;
-			         		totalValor += element.valorCompraIva;
-			          		logger.error('Actualiza '+data);
-			         		logger.error(numeroActivos + ' ' + totalValor);
-			            }).catch(err => res.status(501).json({error: err}));
- */  					}else if (element.estado == "X"){
+  					}else if (element.estado == "X"){
 					//Registro a eliminar
 						try{
 							const eje = await ActaDetalle.destroy({where: {id: element.id}, transaction: transaction})
@@ -349,7 +200,6 @@ module.exports = {
 	async aprueba(req, res){
 		 
 		try{
-			//logger.error(req);
 		   let query = "begin EAFPK_ACTA_CAMBIOESTADO.EAFPR_APRUEBA("+req.params.id+",'"+req.query.usuario+"'); end;";
 		   const[result, metadata] = await db.sequelize.query(query);
 		   res.status(200).json({success:true, message: "Validacion", respuesta: metadata});	
@@ -357,11 +207,6 @@ module.exports = {
 		   logger.error(err);
 		   res.status(501).json({error: err.stack});
 		}
-	
-
    }
-
-
-
 }
